core: AtomDecomposition: Hide implementation
