+++ src/core/AtomDecomposition.hpp
@@ -27,12 +27,8 @@
 
 #include <utils/Vector.hpp>
 
-#include <boost/mpi/collectives/all_to_all.hpp>
 #include <boost/mpi/communicator.hpp>
-#include <boost/serialization/vector.hpp>
 
-#include <cassert>
-#include <limits>
 #include <vector>
 
 /**
@@ -41,7 +37,7 @@
  * This implements a distributed particle storage
  * by just evenly distributing the particles over
  * all part-taking nodes. Pairs are found by just
- * considering all pairs indepent of logical or
+ * considering all pairs independent of logical or
  * physical location, it has therefor quadratic time
  * complexity in the number of particles.
  *
@@ -58,48 +54,10 @@ class AtomDecomposition {
   GhostCommunicator m_collect_ghost_force_comm;
 
 public:
-  explicit AtomDecomposition(boost::mpi::communicator const &comm)
-      : comm(comm), cells(comm.size()) {
-    /* create communicators */
-    configure_comms();
-    /* configure neighbor relations */
-    configure_neighbors();
-    /* fill local and ghost cell lists */
-    mark_cells();
-  }
+  explicit AtomDecomposition(boost::mpi::communicator const &comm);
 
   void resort(bool global_flag, ParticleList &displaced_parts,
-              std::vector<Cell *> &modified_cells) {
-    /* Local updates are a NoOp for this decomposition. */
-    if (not global_flag) {
-      assert(displaced_parts.empty());
-      return;
-    }
-
-    /* Sort displaced particles by the node they belong to. */
-    std::vector<std::vector<Particle>> send_buf(comm.size());
-    for (auto &p : displaced_parts) {
-      auto const target_node = id_to_rank(p.identity());
-      send_buf.at(target_node).emplace_back(std::move(p));
-    }
-    displaced_parts.clear();
-
-    /* Exchange particles */
-    std::vector<std::vector<Particle>> recv_buf(comm.size());
-    boost::mpi::all_to_all(comm, send_buf, recv_buf);
-
-    if (std::any_of(recv_buf.begin(), recv_buf.end(),
-                    [](auto const &buf) { return not buf.empty(); })) {
-      modified_cells.push_back(std::addressof(local()));
-    }
-
-    /* Add new particles belonging to this node */
-    for (auto &parts : recv_buf) {
-      for (auto &p : parts) {
-        local().particles().insert(std::move(p));
-      }
-    }
-  }
+              std::vector<Cell *> &modified_cells);
 
   GhostCommunicator const &exchange_ghosts_comm() const {
     return m_exchange_ghosts_comm;
@@ -111,20 +69,23 @@ public:
   std::vector<Cell *> const &local_cells() const { return m_local_cells; }
   std::vector<Cell *> const &ghost_cells() const { return m_ghost_cells; }
 
-  Cell *particle_to_cell(Particle const &p) { return id_to_cell(p.identity()); }
-
-  Utils::Vector3d max_range() const {
-    return Utils::Vector3d::broadcast(std::numeric_limits<double>::infinity());
-  }
-
-private:
   /**
    * @brief Determine which cell a particle id belongs to.
    *
-   * Since there is only one cell this is trivial.
+   * Since there is only one local cell this is trivial.
    *
-   * @param id Particle id to find cell for.
+   * @param p Particle to find cell for.
    * @return Pointer to cell or nullptr if not local.
+   */
+  Cell *particle_to_cell(Particle const &p) { return id_to_cell(p.identity()); }
+
+  Utils::Vector3d max_range() const;
+
+private:
+  /**
+   * @brief Find cell for id.
+   * @param id to find cell for.
+   * @return Cell for id.
    */
   Cell *id_to_cell(int id) {
     return (id_to_rank(id) == comm.rank()) ? std::addressof(local()) : nullptr;
@@ -135,82 +96,18 @@ private:
    */
   Cell &local() { return cells.at(comm.rank()); }
 
-  void configure_neighbors() {
-    std::vector<Cell *> red_neighbors;
-    std::vector<Cell *> black_neighbors;
-
-    /* distribute force calculation work  */
-    for (int n = 0; n < comm.size(); n++) {
-      auto const diff = n - comm.rank();
-      /* simple load balancing formula. Basically diff % n, where n >=
-         ad.comm.size(), n odd. The node itself is also left out, as it is
-         treated differently */
-      if (diff == 0) {
-        continue;
-      }
-
-      if (((diff > 0 && (diff % 2) == 0) || (diff < 0 && ((-diff) % 2) == 1))) {
-        red_neighbors.push_back(&cells.at(n));
-      } else {
-        black_neighbors.push_back(&cells.at(n));
-      }
-    }
-
-    local().m_neighbors = Neighbors<Cell *>(red_neighbors, black_neighbors);
-  }
-
-  GhostCommunicator prepare_comm() {
-    /* no need for comm for only 1 node */
-    if (comm.size() == 1) {
-      return GhostCommunicator{comm, 0};
-    }
-
-    auto ghost_comm = GhostCommunicator{comm, static_cast<size_t>(comm.size())};
-    /* every node has its dedicated comm step */
-    for (int n = 0; n < comm.size(); n++) {
-      ghost_comm.communications[n].part_lists.resize(1);
-      ghost_comm.communications[n].part_lists[0] = &(cells.at(n).particles());
-      ghost_comm.communications[n].node = n;
-    }
-
-    return ghost_comm;
-  }
+  void configure_neighbors();
+  GhostCommunicator prepare_comm();
 
   /**
    * @brief Setup ghost communicators.
    */
-  void configure_comms() {
-    m_exchange_ghosts_comm = prepare_comm();
-    m_collect_ghost_force_comm = prepare_comm();
+  void configure_comms();
 
-    if (comm.size() > 1) {
-      for (int n = 0; n < comm.size(); n++) {
-        /* use the prefetched send buffers. Node 0 transmits first and never
-         * prefetches. */
-        if (comm.rank() == 0 || comm.rank() != n) {
-          m_exchange_ghosts_comm.communications[n].type = GHOST_BCST;
-        } else {
-          m_exchange_ghosts_comm.communications[n].type =
-              GHOST_BCST | GHOST_PREFETCH;
-        }
-        m_collect_ghost_force_comm.communications[n].type = GHOST_RDCE;
-      }
-      /* first round: all nodes except the first one prefetch their send data */
-      if (comm.rank() != 0) {
-        m_exchange_ghosts_comm.communications[0].type |= GHOST_PREFETCH;
-      }
-    }
-  }
-
-  void mark_cells() {
-    m_local_cells.resize(1, std::addressof(local()));
-    m_ghost_cells.clear();
-    for (int n = 0; n < comm.size(); n++) {
-      if (n != comm.rank()) {
-        m_ghost_cells.push_back(std::addressof(cells.at(n)));
-      }
-    }
-  }
+  /**
+   * @brief Mark local and ghost cells.
+   */
+  void mark_cells();
 
   /**
    * @brief Determine which rank owns a particle id.
